backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langchain>=1.2.0
langchain-anthropic>=1.4.0
langchain-community>=0.4.0
langchain-chroma>=0.2.0
langchain-classic>=1.0.0
langchain-text-splitters>=1.1.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0